Add files via upload
The MainTest.py file contains the primary code of the project. AboutMe.py is a test file containing code to create a Label in a second window. InputSign.py contains code to create another window in case of bad user input. 

I consider to take the class Functions and paste the code to a separate file, adding AboutMe and InputSign to the Functions class.